requirements.txt
blinker==1.9.0
click==8.1.8
DateTime==5.5
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
setuptools==78.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
zope.interface==7.2
